feat: save Gemini prediction in price_tag_predictions table
