+++ open_prices/proofs/constants.py
@@ -18,12 +18,16 @@ TYPE_ALLOW_ANY_USER_PRICE_ADD_LIST = [TYPE_PRICE_TAG]
 PROOF_PREDICTION_OBJECT_DETECTION_TYPE = "OBJECT_DETECTION"
 PROOF_PREDICTION_CLASSIFICATION_TYPE = "CLASSIFICATION"
 PROOF_PREDICTION_RECEIPT_EXTRACTION_TYPE = "RECEIPT_EXTRACTION"
-PROOF_PREDICTION_PRICE_TAG_EXTRACTION_TYPE = "PRICE_TAG_EXTRACTION"
 PROOF_PREDICTION_LIST = [
     PROOF_PREDICTION_OBJECT_DETECTION_TYPE,
     PROOF_PREDICTION_CLASSIFICATION_TYPE,
     PROOF_PREDICTION_RECEIPT_EXTRACTION_TYPE,
-    PROOF_PREDICTION_PRICE_TAG_EXTRACTION_TYPE,
 ]
 
 PROOF_TYPE_CHOICES = [(key, key) for key in PROOF_PREDICTION_LIST]
+
+PRICE_TAG_EXTRACTION_TYPE = "PRICE_TAG_EXTRACTION"
+
+PRICE_TAG_PREDICTION_TYPE_CHOICES = [
+    (PRICE_TAG_EXTRACTION_TYPE, PRICE_TAG_EXTRACTION_TYPE)
+]
